pyproject.toml
[project]
name = "django-reversion-compare"
dynamic = ["version"]
description = "Add compare view to django-reversion for comparing two versions of a reversion model. "
readme = "README.md"
authors = [
    {name = 'Jens Diemer', email = 'django-reversion-compare@jensdiemer.de'}
]
requires-python = ">=3.9,<4" # Keep Python 3.9 until Yunohost contains a newer Python Version ;)
dependencies = [
    "django",
    "django-reversion",
]
[project.optional-dependencies]
dev = [
    "diff-match-patch",
    "bx_django_utils",   # https://github.com/boxine/bx_django_utils
    "django-debug-toolbar",  # https://github.com/jazzband/django-debug-toolbar/
    "django-countries",  # https://github.com/SmileyChris/django-countries

    "colorlog",  # https://github.com/borntyping/python-colorlog
    "gunicorn",  # https://github.com/benoimyproject.wsgitc/gunicorn

    "manage_django_project>=0.3.0",  # https://github.com/jedie/manage_django_project
    "tblib",   # https://github.com/ionelmc/python-tblib
    "pip-tools",  # https://github.com/jazzband/pip-tools/
    "tox",  # https://github.com/tox-dev/tox
    "coverage",  # https://github.com/nedbat/coveragepy
    "autopep8",  # https://github.com/hhatto/autopep8
    "pyupgrade",  # https://github.com/asottile/pyupgrade
    "flake8",  # https://github.com/pycqa/flake8
    "pyflakes",  # https://github.com/PyCQA/pyflakes
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "safety",  # https://github.com/pyupio/safety
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine

    # https://github.com/akaihola/darker
    # https://github.com/ikamensh/flynt
    # https://github.com/pycqa/isort
    # https://github.com/pygments/pygments
    "darker[flynt, isort, color]",

    "tomli",  # https://github.com/hukkin/tomli
    # tomli only needed for Python <3.11, but see bug:
    # https://github.com/pypa/pip/issues/9644#issuecomment-1456583402
    #"tomli;python_version<\"3.11\"",  # https://github.com/hukkin/tomli

    # Work-a-round for:
    # https://github.com/jazzband/pip-tools/issues/994#issuecomment-1321226661
    "typing-extensions>=3.10",

    "model_bakery",  # https://github.com/model-bakers/model_bakery
    "requests-mock",
    "django-override-storage",  # https://github.com/danifus/django-override-storage
]

[project.urls]
Documentation = "https://github.com/jedie/django-reversion-compare"
Source = "https://github.com/jedie/django-reversion-compare"


[project.scripts]
reversion_compare_project = "reversion_compare_project.__main__:main"

[manage_django_project]
module_name="reversion_compare_project"

# Django settings used for all commands except test/coverage/tox:
local_settings='reversion_compare_project.settings.local'

# Django settings used for test/coverage/tox commands:
test_settings='reversion_compare_project.settings.tests'


[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=7.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["reversion_compare*", "reversion_compare_project*"]

[tool.setuptools.dynamic]
version = {attr = "reversion_compare.__version__"}


[tool.darker]
src = ['.']
revision = "origin/main..."
line_length = 119
verbose = true
color = true
skip_string_normalization = true
diff = false
check = false
stdout = false
isort = true
flynt = true
lint = [
    "flake8",
]
log_level = "INFO"


[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files/#pyprojecttoml-preferred-format
atomic=true
profile='black'
skip_glob=['.*', '*/htmlcov/*','*/migrations/*']
known_first_party=['reversion_compare']
line_length=119
lines_after_imports=2


[tool.coverage.run]
branch = true
parallel = true
concurrency = ["multiprocessing"]
source = ['.']
command_line = '-m reversion_compare_project test --shuffle --parallel --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*', '*/migrations/*']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.tox]  # https://tox.wiki/en/latest/config.html#pyproject-toml
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py{311,310,39}
skip_missing_interpreters = True

[testenv]
passenv = *
skip_install = true
commands_pre =
    pip install -U pip-tools
    pip-sync requirements.dev.txt
commands =
    {envpython} -m coverage run --context='{envname}'
    {envpython} -m coverage combine --append --keep
    {envpython} -m coverage xml
    {envpython} -m coverage report
"""


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests', 'migrations']


[manageprojects] # https://github.com/jedie/manageprojects
initial_revision = "02d64b0"
initial_date = 2023-04-06T11:01:20+02:00
cookiecutter_template = "https://github.com/jedie/cookiecutter_templates/"
cookiecutter_directory = "managed-django-project"

[manageprojects.cookiecutter_context.cookiecutter]
full_name = "Jens Diemer"
github_username = "jedie"
author_email = "django-reversion-compare@jensdiemer.de"
package_name = "reversion_compare"
package_version = "0.15.0"
package_description = "Add compare view to django-reversion for comparing two versions of a reversion model. "
package_url = "https://github.com/jedie/django-reversion-compare"
license = "GPL-3.0-or-later"
_template = "https://github.com/jedie/cookiecutter_templates/"
